WIP: Bug 1811076: Integrate content analysis client
Bug 1811076: Primitive working UI, integrated with downloads

UI just uses window.alert to open an unstyled dialog.  I suspect that nearly all of the UI will end up as notifications so I haven't styled anything.  The result message will probably be a dialog (in the event that the result is not approval).

WIP Bug 1811076: notifications