Bug 1811076: Part 2 - Create nsIContentAnalysis and contentAnalysisViews.js r=cmartin
Add the protocol buffer and client side of the SDK to the Gecko build.  This
includes prefs to enable/disable/configure content analysis, and a JS class
for UI elements that will be shown in response to DLP operations. In addition to the pref, content analysis requires a command-line argument "-allow-content-analysis" to be turned on; this is for security reasons until we make sure that the pref can only be turned on by enterprise policies.

Differential Revision: https://phabricator.services.mozilla.com/D189568