Bug 1811076: Part 5 - Set up methods to consult content analysis for clipboard content r=nika
When a content analysis agent is connected, pastes will be checked by the CA
agent while tab input is blocked.  The CA connection is in the main
process and they must block clipboard operation being done in the content
process main thread, so some operations will SpinEventLoopUntil the
analysis result is received.

Differential Revision: https://phabricator.services.mozilla.com/D189571